Isolated multicast code into separate file
