+++ lib/virtual-net/src/host.rs
@@ -1,4 +1,7 @@
 #![allow(unused_variables)]
+use crate::multicast::{
+    LocalUdpSocketInterestHandler, LocalUdpSocketShared, MulticastCoordinator, MulticastKey,
+};
 use crate::ruleset::{Direction, Ruleset};
 #[allow(unused_imports)]
 use crate::{
@@ -8,7 +11,7 @@ use crate::{
 };
 use crate::{VirtualIoSource, io_err_into_net_error};
 use bytes::{Buf, BytesMut};
-use std::collections::{HashMap, VecDeque};
+use std::collections::VecDeque;
 use std::io::{self, Read, Write};
 use std::mem::MaybeUninit;
 use std::net::{IpAddr, Ipv4Addr, Ipv6Addr, Shutdown, SocketAddr};
@@ -20,9 +23,8 @@ use std::os::fd::RawFd;
 use std::os::windows::io::AsRawSocket;
 
 use std::sync::atomic::{AtomicU64, Ordering};
-use std::sync::{Arc, Mutex, Weak};
+use std::sync::{Arc, Mutex};
 use std::task::Poll;
-use std::task::Waker;
 use std::time::Duration;
 use tokio::runtime::Handle;
 #[allow(unused_imports, dead_code)]
@@ -71,253 +73,6 @@ impl Drop for LocalNetworking {
 impl Default for LocalNetworking {
     fn default() -> Self {
         Self::new()
-    }
-}
-
-const MULTICAST_RING_CAPACITY: usize = 1024;
-
-#[derive(Clone, Copy, Debug, Hash, PartialEq, Eq)]
-struct MulticastKey {
-    addr: IpAddr,
-    port: u16,
-}
-
-impl MulticastKey {
-    fn from_socket_addr(addr: SocketAddr) -> Option<Self> {
-        match addr.ip() {
-            IpAddr::V4(ip) if ip.is_multicast() => Some(Self {
-                addr: IpAddr::V4(ip),
-                port: addr.port(),
-            }),
-            IpAddr::V6(ip) if ip.is_multicast() => Some(Self {
-                addr: IpAddr::V6(ip),
-                port: addr.port(),
-            }),
-            _ => None,
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-struct MulticastPacket {
-    seq: u64,
-    data: Arc<[u8]>,
-    from: SocketAddr,
-}
-
-#[derive(Debug, Default)]
-struct MulticastGroup {
-    next_seq: u64,
-    packets: VecDeque<MulticastPacket>,
-    members: HashMap<u64, Weak<LocalUdpSocketShared>>,
-}
-
-#[derive(Debug, Default)]
-struct MulticastCoordinator {
-    groups: HashMap<MulticastKey, MulticastGroup>,
-}
-
-impl MulticastCoordinator {
-    fn join(&mut self, socket: &Arc<LocalUdpSocketShared>, key: MulticastKey) {
-        let group = self.groups.entry(key).or_default();
-        group.members.insert(socket.id, Arc::downgrade(socket));
-        socket
-            .multicast_reads
-            .lock()
-            .unwrap()
-            .insert(key, group.next_seq);
-    }
-
-    fn leave(&mut self, socket: &Arc<LocalUdpSocketShared>, key: MulticastKey) {
-        if let Some(group) = self.groups.get_mut(&key) {
-            group.members.remove(&socket.id);
-            if group.members.is_empty() {
-                self.groups.remove(&key);
-            }
-        }
-        socket.multicast_reads.lock().unwrap().remove(&key);
-    }
-
-    fn send(
-        &mut self,
-        sender: &Arc<LocalUdpSocketShared>,
-        data: &[u8],
-        from: SocketAddr,
-        to: SocketAddr,
-    ) -> Vec<Arc<LocalUdpSocketShared>> {
-        let Some(key) = MulticastKey::from_socket_addr(to) else {
-            return Vec::new();
-        };
-        let Some(group) = self.groups.get_mut(&key) else {
-            return Vec::new();
-        };
-
-        let packet = MulticastPacket {
-            seq: group.next_seq,
-            data: Arc::from(data),
-            from,
-        };
-        group.next_seq = group.next_seq.wrapping_add(1);
-        group.packets.push_back(packet);
-        while group.packets.len() > MULTICAST_RING_CAPACITY {
-            group.packets.pop_front();
-        }
-
-        let mut stale = Vec::new();
-        let mut subscribers = Vec::new();
-        for (&id, member) in &group.members {
-            let Some(member) = member.upgrade() else {
-                stale.push(id);
-                continue;
-            };
-            if Arc::ptr_eq(sender, &member) && !member.multicast_loop_for(key.addr) {
-                if let Some(cursor) = member.multicast_reads.lock().unwrap().get_mut(&key) {
-                    *cursor = group.next_seq;
-                }
-                continue;
-            }
-            subscribers.push(member);
-        }
-        for id in stale {
-            group.members.remove(&id);
-        }
-        subscribers
-    }
-
-    fn next_packet_len(&mut self, socket: &Arc<LocalUdpSocketShared>) -> Option<usize> {
-        let mut reads = socket.multicast_reads.lock().unwrap();
-        for (key, cursor) in reads.iter_mut() {
-            let Some(group) = self.groups.get(key) else {
-                continue;
-            };
-            if !group.members.contains_key(&socket.id) {
-                continue;
-            }
-            let Some(front) = group.packets.front() else {
-                continue;
-            };
-            if *cursor < front.seq {
-                *cursor = front.seq;
-            }
-            if let Some(packet) = group.packets.iter().find(|packet| packet.seq >= *cursor) {
-                return Some(packet.data.len());
-            }
-        }
-        None
-    }
-
-    fn recv(
-        &mut self,
-        socket: &Arc<LocalUdpSocketShared>,
-        buf: &mut [MaybeUninit<u8>],
-        peek: bool,
-    ) -> Result<(usize, SocketAddr)> {
-        let mut reads = socket.multicast_reads.lock().unwrap();
-        for (key, cursor) in reads.iter_mut() {
-            let Some(group) = self.groups.get(key) else {
-                continue;
-            };
-            if !group.members.contains_key(&socket.id) {
-                continue;
-            }
-            let Some(front) = group.packets.front() else {
-                continue;
-            };
-            if *cursor < front.seq {
-                *cursor = front.seq;
-            }
-            let Some(packet) = group.packets.iter().find(|packet| packet.seq >= *cursor) else {
-                continue;
-            };
-
-            let amt = buf.len().min(packet.data.len());
-            let buf: &mut [u8] = unsafe { std::mem::transmute(buf) };
-            buf[..amt].copy_from_slice(&packet.data[..amt]);
-            if !peek {
-                *cursor = packet.seq.wrapping_add(1);
-            }
-            return Ok((amt, packet.from));
-        }
-        Err(NetworkError::WouldBlock)
-    }
-}
-
-struct LocalUdpSocketShared {
-    id: u64,
-    multicast_reads: Mutex<HashMap<MulticastKey, u64>>,
-    multicast_loop_v4: Mutex<bool>,
-    multicast_loop_v6: Mutex<bool>,
-    read_wakers: Mutex<Vec<Waker>>,
-    handler: Mutex<Option<Box<dyn InterestHandler + Send + Sync>>>,
-}
-
-impl std::fmt::Debug for LocalUdpSocketShared {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("LocalUdpSocketShared")
-            .field("id", &self.id)
-            .finish_non_exhaustive()
-    }
-}
-
-impl LocalUdpSocketShared {
-    fn new(id: u64) -> Self {
-        Self {
-            id,
-            multicast_reads: Default::default(),
-            multicast_loop_v4: Mutex::new(true),
-            multicast_loop_v6: Mutex::new(true),
-            read_wakers: Default::default(),
-            handler: Default::default(),
-        }
-    }
-
-    fn multicast_loop_for(&self, addr: IpAddr) -> bool {
-        match addr {
-            IpAddr::V4(_) => *self.multicast_loop_v4.lock().unwrap(),
-            IpAddr::V6(_) => *self.multicast_loop_v6.lock().unwrap(),
-        }
-    }
-
-    fn notify_readable(&self) {
-        for waker in self.read_wakers.lock().unwrap().drain(..) {
-            waker.wake();
-        }
-        if let Some(handler) = self.handler.lock().unwrap().as_mut() {
-            handler.push_interest(InterestType::Readable);
-        }
-    }
-}
-
-#[derive(Debug)]
-struct LocalUdpSocketInterestHandler {
-    shared: Arc<LocalUdpSocketShared>,
-}
-
-impl InterestHandler for LocalUdpSocketInterestHandler {
-    fn push_interest(&mut self, interest: InterestType) {
-        if let Some(handler) = self.shared.handler.lock().unwrap().as_mut() {
-            handler.push_interest(interest);
-        }
-    }
-
-    fn pop_interest(&mut self, interest: InterestType) -> bool {
-        self.shared
-            .handler
-            .lock()
-            .unwrap()
-            .as_mut()
-            .map(|handler| handler.pop_interest(interest))
-            .unwrap_or(false)
-    }
-
-    fn has_interest(&self, interest: InterestType) -> bool {
-        self.shared
-            .handler
-            .lock()
-            .unwrap()
-            .as_ref()
-            .map(|handler| handler.has_interest(interest))
-            .unwrap_or(false)
     }
 }
 
@@ -1068,14 +823,7 @@ pub struct LocalUdpSocket {
 
 impl Drop for LocalUdpSocket {
     fn drop(&mut self) {
-        let keys: Vec<_> = self
-            .shared
-            .multicast_reads
-            .lock()
-            .unwrap()
-            .keys()
-            .copied()
-            .collect();
+        let keys = self.shared.joined_keys();
         let mut multicast = self.multicast.lock().unwrap();
         for key in keys {
             multicast.leave(&self.shared, key);
@@ -1095,25 +843,25 @@ impl VirtualUdpSocket for LocalUdpSocket {
     }
 
     fn set_multicast_loop_v4(&mut self, val: bool) -> Result<()> {
-        *self.shared.multicast_loop_v4.lock().unwrap() = val;
+        self.shared.set_multicast_loop_v4(val);
         self.socket
             .set_multicast_loop_v4(val)
             .map_err(io_err_into_net_error)
     }
 
     fn multicast_loop_v4(&self) -> Result<bool> {
-        Ok(*self.shared.multicast_loop_v4.lock().unwrap())
+        Ok(self.shared.multicast_loop_v4())
     }
 
     fn set_multicast_loop_v6(&mut self, val: bool) -> Result<()> {
-        *self.shared.multicast_loop_v6.lock().unwrap() = val;
+        self.shared.set_multicast_loop_v6(val);
         self.socket
             .set_multicast_loop_v6(val)
             .map_err(io_err_into_net_error)
     }
 
     fn multicast_loop_v6(&self) -> Result<bool> {
-        Ok(*self.shared.multicast_loop_v6.lock().unwrap())
+        Ok(self.shared.multicast_loop_v6())
     }
 
     fn set_multicast_ttl_v4(&mut self, ttl: u32) -> Result<()> {
@@ -1130,49 +878,37 @@ impl VirtualUdpSocket for LocalUdpSocket {
 
     fn join_multicast_v4(&mut self, multiaddr: Ipv4Addr, iface: Ipv4Addr) -> Result<()> {
         let port = self.addr_local()?.port();
-        self.multicast.lock().unwrap().join(
-            &self.shared,
-            MulticastKey {
-                addr: IpAddr::V4(multiaddr),
-                port,
-            },
-        );
+        self.multicast
+            .lock()
+            .unwrap()
+            .join(&self.shared, MulticastKey::new(IpAddr::V4(multiaddr), port));
         Ok(())
     }
 
     fn leave_multicast_v4(&mut self, multiaddr: Ipv4Addr, iface: Ipv4Addr) -> Result<()> {
         let port = self.addr_local()?.port();
-        self.multicast.lock().unwrap().leave(
-            &self.shared,
-            MulticastKey {
-                addr: IpAddr::V4(multiaddr),
-                port,
-            },
-        );
+        self.multicast
+            .lock()
+            .unwrap()
+            .leave(&self.shared, MulticastKey::new(IpAddr::V4(multiaddr), port));
         Ok(())
     }
 
     fn join_multicast_v6(&mut self, multiaddr: Ipv6Addr, iface: u32) -> Result<()> {
         let port = self.addr_local()?.port();
-        self.multicast.lock().unwrap().join(
-            &self.shared,
-            MulticastKey {
-                addr: IpAddr::V6(multiaddr),
-                port,
-            },
-        );
+        self.multicast
+            .lock()
+            .unwrap()
+            .join(&self.shared, MulticastKey::new(IpAddr::V6(multiaddr), port));
         Ok(())
     }
 
     fn leave_multicast_v6(&mut self, multiaddr: Ipv6Addr, iface: u32) -> Result<()> {
         let port = self.addr_local()?.port();
-        self.multicast.lock().unwrap().leave(
-            &self.shared,
-            MulticastKey {
-                addr: IpAddr::V6(multiaddr),
-                port,
-            },
-        );
+        self.multicast
+            .lock()
+            .unwrap()
+            .leave(&self.shared, MulticastKey::new(IpAddr::V6(multiaddr), port));
         Ok(())
     }
 
@@ -1270,14 +1006,12 @@ impl VirtualSocket for LocalUdpSocket {
     }
 
     fn set_handler(&mut self, handler: Box<dyn InterestHandler + Send + Sync>) -> Result<()> {
-        *self.shared.handler.lock().unwrap() = Some(handler);
+        self.shared.set_handler(Some(handler));
 
         if let HandlerGuardState::ExternalHandler(_) = &mut self.handler_guard {
             return Ok(());
         }
-        let handler = Box::new(LocalUdpSocketInterestHandler {
-            shared: self.shared.clone(),
-        });
+        let handler = Box::new(LocalUdpSocketInterestHandler::new(self.shared.clone()));
 
         let guard = InterestGuard::new(
             &self.selector,
@@ -1321,8 +1055,8 @@ impl LocalUdpSocket {
 
 impl VirtualIoSource for LocalUdpSocket {
     fn remove_handler(&mut self) {
-        *self.shared.handler.lock().unwrap() = None;
-        self.shared.read_wakers.lock().unwrap().clear();
+        self.shared.set_handler(None);
+        self.shared.clear_read_wakers();
         let mut guard = HandlerGuardState::None;
         std::mem::swap(&mut guard, &mut self.handler_guard);
         match guard {
@@ -1349,12 +1083,7 @@ impl VirtualIoSource for LocalUdpSocket {
         let map = state_as_waker_map(state, selector, socket).map_err(io_err_into_net_error)?;
         map.pop(InterestType::Readable);
         map.add(InterestType::Readable, cx.waker());
-        {
-            let mut wakers = self.shared.read_wakers.lock().unwrap();
-            if !wakers.iter().any(|waker| waker.will_wake(cx.waker())) {
-                wakers.push(cx.waker().clone());
-            }
-        }
+        self.shared.add_read_waker(cx.waker());
 
         match self.recv_into_backlog() {
             Ok(()) => {
